api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
python-multipart==0.0.6
asyncpg==0.30.0
jinja2==3.1.2
pydantic-settings==2.1.0
alembic==1.13.1
greenlet==3.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4